Use env variables, move functions related with users to the UserHelper.
